feat: totalElement를 응답하도록 수정

+++ src/main/java/com/codeit/deokhugam/domain/entity/Book.java
@@ -36,7 +36,7 @@ public class Book extends BaseUpdatableEntity {
   @Column(name = "thumbnail_url", length = 500)
   private String thumbnailUrl;
   @Column(name = "review_count", nullable = false)
-  private int reviewCount = 0;
+  private Long reviewCount = 0L;
   @Column(name = "rating", nullable = false, precision = 3, scale = 2)
   private BigDecimal rating = BigDecimal.ZERO;
   @Column(name = "deleted", nullable = false)
